refactor: split models.py into models/ package and broaden Anthropic detection
- Split 994-line models.py into models/{constants,upstream,registry}.py
- Add ANTHROPIC_CODENAMES for detecting models by codename infix
  (sonnet, opus, haiku, fable) in addition to claude prefix
- Unify three duplicate _classify_model_family methods into one
  module-level classify_model_family()
- Remove dead code: utils/models.py (never imported)
- Re-export public API from models/__init__.py — no caller changes needed
